// client/src/types/Page.ts
type PageData = {
  name: string;
  section1Title: string;
  section2Title: string;
  section3Title: string;
  section4Title: string;
  section5Title: string;
  section1Link: string;
  section3Link: string;
  section4Link: string;
  section5Link: string;
  section2Content: string;
  section3Content: string;
  img: string;
};

export default PageData;